requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Deprecated==1.2.14
Django==5.0.6
django-prometheus==2.3.1
djangorestframework==3.15.1
googleapis-common-protos==1.63.0
grpcio==1.64.0
gunicorn==22.0.0
idna==3.7
importlib-metadata==7.0.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-django==0.45b0
opentelemetry-instrumentation-logging==0.45b0
opentelemetry-instrumentation-wsgi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
packaging==24.0
prometheus_client==0.20.0
protobuf==4.25.3
requests==2.32.2
setuptools==70.0.0
sqlparse==0.5.0
typing_extensions==4.12.0
urllib3==2.2.1
wrapt==1.16.0
zipp==3.18.2
